implemented protobuf and optimized data sent over network

+++ validate/generate.go
@@ -8,9 +8,9 @@ import (
 // PlayerID Creates and Validates ID to be unique
 
 //GenerateID a unique ID beteen 0 and 100
-func GenerateID(size int) int {
+func GenerateID(size int) uint32 {
 	timestamp := time.Now().UnixNano()
 	rand.Seed(timestamp)
-	id := rand.Intn(size)
+	id := uint32(rand.Intn(size))
 	return id
 }
